services/agentic-hr/ingestion/requirements.txt
docling<2.80
psycopg2-binary
pgvector
anthropic
sentence-transformers
python-dotenv
numpy<2
transformers<5
